add authentication
add user for appointments and projects

+++ src/main/java/at/gartner/calendar/controller/CalendarController.java
@@ -14,6 +14,7 @@ import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.format.annotation.DateTimeFormat.ISO;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -21,12 +22,16 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import at.gartner.calendar.CalendarUserDetailsService;
+import at.gartner.calendar.CalendarUserPrincipal;
 import at.gartner.calendar.data.entity.ActivityType;
 import at.gartner.calendar.data.entity.Appointment;
 import at.gartner.calendar.data.entity.Project;
+import at.gartner.calendar.data.entity.User;
 import at.gartner.calendar.data.repository.ActivityTypeRepository;
 import at.gartner.calendar.data.repository.AppointmentRepository;
 import at.gartner.calendar.data.repository.ProjectRepository;
+import at.gartner.calendar.data.repository.UserRepository;
 import at.gartner.calendar.data.viewmodel.SchedulerAppointment;
 import at.gartner.calendar.data.viewmodel.WeeklySummaryItem;
 
@@ -40,50 +45,74 @@ public class CalendarController {
 	private ActivityTypeRepository activityTypeRepository;
 	@Autowired
 	private AppointmentRepository appointmentRepository;
+	@Autowired
+	private UserRepository userRepository;
 
 	@RequestMapping(value = "/loadAppointmentsForWeek/{startDate}")
-	public Iterable<Appointment> LoadAppointmentsForWeek(@PathVariable(name = "startDate") @DateTimeFormat(iso = ISO.DATE_TIME) LocalDateTime startDate)
+	public Iterable<Appointment> LoadAppointmentsForWeek(Authentication authentication, @PathVariable(name = "startDate") @DateTimeFormat(iso = ISO.DATE_TIME) LocalDateTime startDate)
 	{
-		//TODO: only find by week (for simplicity just date +- 7 days)
-		Iterable<Appointment> appointments = appointmentRepository.findAll();
+		CalendarUserPrincipal p = (CalendarUserPrincipal)authentication.getPrincipal();
+		
+		//TODO: only find by week (for simplicity just date +- 7 days)		
+		Iterable<Appointment> appointments = appointmentRepository.findByUserId(p.getUserId());
 		return appointments;
 	}
 		
 	@RequestMapping(value = "/appointment/insert", method = RequestMethod.POST)
-	public ResponseEntity<SchedulerAppointment> InsertAppointment(@RequestBody SchedulerAppointment appointment)
+	public ResponseEntity<SchedulerAppointment> InsertAppointment(Authentication authentication, @RequestBody SchedulerAppointment appointment)
 	{
+		CalendarUserPrincipal p = (CalendarUserPrincipal)authentication.getPrincipal();
+		User u = userRepository.findById(p.getUserId()).orElse(null);
+
 		Project project = projectRepository.findById(appointment.getProjectId()).orElse(null);
 		ActivityType activityType = activityTypeRepository.findById(appointment.getActivityTypeId()).orElse(null);
-		Appointment data = new Appointment(appointment.getStartDate(), appointment.getEndDate(), appointment.getSubject(), appointment.getBody());
+		Appointment data = new Appointment(appointment.getStartDate(), appointment.getEndDate(), appointment.getSubject(), appointment.getBody());		
 		data.setProject(project);
 		data.setActivityType(activityType);
+		data.setUser(u);
 		appointmentRepository.save(data);
         return new ResponseEntity<SchedulerAppointment>(appointment, HttpStatus.OK);
 	}
 	
 	@RequestMapping(value = "/appointment/update/{id}", method = RequestMethod.POST)
-	public ResponseEntity<SchedulerAppointment> UpdateAppointment(@PathVariable(name = "id") Long id, @RequestBody SchedulerAppointment appointment)
+	public ResponseEntity<SchedulerAppointment> UpdateAppointment(Authentication authentication, @PathVariable(name = "id") Long id, @RequestBody SchedulerAppointment appointment)
 	{
+		CalendarUserPrincipal p = (CalendarUserPrincipal)authentication.getPrincipal();
+		User u = userRepository.findById(p.getUserId()).orElse(null);
+		
+		//TODO: check if appointment is owned by User u
 		Project project = projectRepository.findById(appointment.getProjectId()).orElse(null);
 		ActivityType activityType = activityTypeRepository.findById(appointment.getActivityTypeId()).orElse(null);
 		Appointment data = new Appointment(appointment.getStartDate(), appointment.getEndDate(), appointment.getSubject(), appointment.getBody());
 		data.setId(appointment.getId());
 		data.setProject(project);
 		data.setActivityType(activityType);
+		data.setUser(u);
 		appointmentRepository.save(data);
         return new ResponseEntity<SchedulerAppointment>(appointment, HttpStatus.OK);
 	}
 	
 	@RequestMapping(value = "/appointment/delete/{id}")
-	public String DeleteAppointment(@PathVariable(name = "id") Long id)
+	public String DeleteAppointment(Authentication authentication, @PathVariable(name = "id") Long id)
 	{
-		appointmentRepository.deleteById(id);
-		return "deleted";
+		CalendarUserPrincipal p = (CalendarUserPrincipal)authentication.getPrincipal();
+		Appointment a = appointmentRepository.findById(id).orElse(null);
+		if (a != null && a.getUserId() == p.getUserId())
+		{
+			appointmentRepository.deleteById(id);
+			return "deleted";	
+		}
+		else
+		{
+			return "failed";
+		}
 	}
 
 	@RequestMapping(value = "/loadWeeklySummary/{startDate}")
-	public List<WeeklySummaryItem> LoadWeeklySummary(@PathVariable(name = "startDate") @DateTimeFormat(iso = ISO.DATE_TIME) LocalDateTime startDate)
+	public List<WeeklySummaryItem> LoadWeeklySummary(Authentication authentication, @PathVariable(name = "startDate") @DateTimeFormat(iso = ISO.DATE_TIME) LocalDateTime startDate)
 	{	
+		CalendarUserPrincipal p = (CalendarUserPrincipal)authentication.getPrincipal();
+
 		//TODO: write custom repository method for fetching only weekly apointments!
 		LocalDateTime monday = startDate;
 	    while (monday.getDayOfWeek() != DayOfWeek.MONDAY)
@@ -97,7 +126,7 @@ public class CalendarController {
 	      sunday = sunday.plusDays(1);
 	    }
 		List<WeeklySummaryItem> items = new ArrayList<>();
-		Iterable<Appointment> appointments = appointmentRepository.findAll();
+		Iterable<Appointment> appointments = appointmentRepository.findByUserId(p.getUserId());
 		for(Appointment a : appointments)
 		{
 			if (a.getStartDate().toLocalDateTime().isAfter(monday) && a.getEndDate().toLocalDateTime().isBefore(sunday))
